#33 relocating source files for maven build
