+++ src/main/java/jdiskmark/BenchmarkOperation.java
