+++ src/main/java/jdiskmark/BenchmarkPanel.java
