+++ src/main/java/jdiskmark/BenchmarkWorker.java
