+++ src/main/java/jdiskmark/DriveAccessChecker.java
