+++ src/main/java/jdiskmark/GcDetector.java
