+++ src/main/java/jdiskmark/OperationTableSelectionListener.java
