+++ src/main/java/jdiskmark/PortalEnableDialog.java
