+++ src/main/java/jdiskmark/Util.java
@@ -1,3 +1,5 @@
+
+
 package jdiskmark;
 
 import java.io.BufferedReader;
