+++ src/main/java/jdiskmark/VersionProvider.java
@@ -1,3 +1,4 @@
+
 package jdiskmark;
 
 import picocli.CommandLine.IVersionProvider;
